+++ influxql/parser.go
@@ -1558,7 +1558,7 @@ func (p *Parser) parseFill() (FillOption, interface{}, error) {
 		p.unscan()
 		return NullFill, nil, nil
 	} else {
-		if lit.Name != "fill" {
+		if strings.ToLower(lit.Name) != "fill" {
 			p.unscan()
 			return NullFill, nil, nil
 		}